refactored architecture for better readability and also fixed all errors associated with refactor
