--- logs/mlsvc.log
@@ -1,15 +1,0 @@
-2025-05-03 01:56:46.480 | WARNING  | model.model_service:load_model:33 - Model not found, training a new model.
-2025-05-03 01:56:46.480 | INFO     | model.preparation:prepare_data:85 - Preparing data for model training...
-2025-05-03 01:56:46.480 | INFO     | model.collection:load_data_from_db:31 - Loading data from SQLite database
-2025-05-03 01:56:46.487 | INFO     | model.preparation:prepare_data:89 - Converting boolean columns to one-hot encoding...
-2025-05-03 01:56:46.490 | INFO     | model.preparation:prepare_data:94 - Converting garden column to int...
-2025-05-03 01:56:46.491 | INFO     | model.preparation:prepare_data:138 - Dropping unnecessary columns...
-2025-05-03 01:56:46.492 | INFO     | model.model:build_model:68 - Training model...
-2025-05-03 01:56:46.492 | INFO     | model.model:train_model:30 - Training model with grid search...
-2025-05-03 01:56:46.492 | DEBUG    | model.model:train_model:31 - Grid search space: {'n_estimators': [100, 200, 300], 'max_depth': [3, 6, 9, 12, 15]}
-2025-05-03 01:57:00.198 | DEBUG    | model.model:train_model:36 - Best model params: {'max_depth': 9, 'n_estimators': 200}
-2025-05-03 01:57:00.205 | INFO     | model.model:build_model:73 - Model score: 0.7484955857156256
-2025-05-03 01:57:00.209 | INFO     | model.model:build_model:77 - Model saved to model/models/rf_db_v2
-2025-05-03 01:57:00.210 | INFO     | model.model_service:load_model:36 - Model exists -> loading model.
-2025-05-03 01:57:00.212 | INFO     | model.model_service:load_model:40 - Model loaded successfully.
-2025-05-03 01:57:00.212 | DEBUG    | model.model_service:predict:46 - Input parameters for prediction: [85, 2015, 2, 2, 1, 20, 1, 1, 0, 0, 1]
