+++ runner.py
@@ -1,6 +1,6 @@
 from loguru import logger
 
-from model.model_service import ModelService
+from src.model.model_service import ModelService
 
 
 @logger.catch
