+++ src/config/.env
@@ -1,9 +1,9 @@
 # model service
-MODEL_PATH=model/models
+MODEL_PATH=src/model/models
 MODEL_NAME=rf_db_v2
 
 # logging
-LOG_PATH=logs
+LOG_PATH=src/logs
 SVC_LOG_NAME=mlsvc.log
 SVC_ROTATION=1 day
 SVC_COMPRESSION=zip
