+++ src/config/config.py
@@ -8,7 +8,7 @@ class Settings(BaseSettings):
     """Settings for the application."""
 
     model_config = SettingsConfigDict(
-        env_file=".env",
+        env_file="src/config/.env",
         env_file_encoding="utf-8",
     )
