+++ src/logs/mlsvc.log
@@ -1,0 +1,3 @@
+2025-05-11 23:16:32.295 | INFO     | src.model.model_service:load_model:36 - Model exists -> loading model.
+2025-05-11 23:16:32.301 | INFO     | src.model.model_service:load_model:40 - Model loaded successfully.
+2025-05-11 23:16:32.301 | DEBUG    | src.model.model_service:predict:46 - Input parameters for prediction: [85, 2015, 2, 2, 1, 20, 1, 1, 0, 0, 1]
